Improve metrics system

+++ rust/otap-dataflow/crates/telemetry/src/attributes.rs
@@ -7,7 +7,7 @@
 use std::borrow::Cow;
 
 /// Immutable attributes for a specific instance of a node.
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone, Default, Hash)]
 pub struct NodeStaticAttrs {
     /// Node unique identifier (in scope of the pipeline).
     pub node_id: Cow<'static, str>,
@@ -17,9 +17,9 @@ pub struct NodeStaticAttrs {
     pub pipeline_id: Cow<'static, str>,
 
     /// Core identifier.
-    pub core_id: u16,
+    pub core_id: usize,
     /// NUMA node identifier.
-    pub numa_node_id: u16,
+    pub numa_node_id: usize,
     /// Unique process identifier.
     pub process_id: u32,
 }
